Code that almost works but doesnt
Using selenium to scrape youtube comments. I've tried everything I can get from google to get this to work, but so far nothing has. This code seems to work for others, but it won't run for me in google colab. I keep getting errors with the chrome webdriver part of the code.